pyproject.toml
[project]
name = "candid-satcam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dot>=0.3.0",
    "dotenv>=0.9.9",
    "env>=0.1.0",
    "geopy>=2.4.1",
    "herbie-data[extras]>=2025.12.0",
    "matplotlib>=3.10.8",
    "pvlib>=0.13.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "timezonefinder>=8.2.0",
]
